Finished implementing react-router --- WIP on refactoring the whole project to react app
