+++ package.json
@@ -9,6 +9,7 @@
     "prop-types": "^15.7.2",
     "react": "16.7.0-alpha.0",
     "react-dom": "16.7.0-alpha.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.1",
     "react-spring": "^7.2.8",
     "react-with-gesture": "^4.0.2"
