+++ src/App.jsx
@@ -1,8 +1,10 @@
-import React, { Component, Fragment } from "react";
-import Deck from "./components/Deck";
-import Table from "./components/Table";
+import React from "react";
+import LoginPage from "./components/LoginPage/LoginPage";
+import DeckPage from "./components/DeckPage/DeckPage";
+import Table from "./components/Table/Table";
 import firebase from "firebase";
-import fire from "./fire";
+
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 // let firebaseConfig = {
 //   // Your web app's Firebase configuration
@@ -18,13 +20,13 @@ import fire from "./fire";
 // firebase.initializeApp(firebaseConfig);
 //let dbFavMovies = firebase.firestore();
 function App() {
-  // let database = firebase.firestore();
-  // // let dbFavMovies = firebase.firestore();
+  let database = firebase.firestore();
+  // let dbFavMovies = firebase.firestore();
 
   // let yourFavoriteMovie = prompt("Enter your favorite movie?")
   // let movieRating = prompt("Who is the movie rating?")
   // let releaseYear = prompt("What is release year?")
-  
+
   // database
   //   .collection("RealTable")
   //   .add({
@@ -42,13 +44,14 @@ function App() {
   //   });
 
   return (
-      <React.Fragment>
-      
-      <h1>MovieM8S</h1>
-      <Deck />
-      <Table />
-      </React.Fragment>
-    
+      <Router>
+        <h1>MovieM8S</h1>
+        <Switch>
+          <Route exact path="/" component={LoginPage}/>
+          <Route exact path="/deck" component={DeckPage}/>
+          <Route exact path="/favorite-movies" component={Table}/>
+        </Switch>
+      </Router>
   );
 }
