+++ src/components/LoginPage/LoginPage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import {Link} from "react-router-dom";
+
+export default function LoginPage(){
+  return(
+    <div>
+      <h1>Login Page</h1>
+      <Link to={"/deck"}>
+        <button>Login</button>
+      </Link>
+    </div>
+  )
+}
