--- src/components/Table.js
@@ -1,29 +1,0 @@
-import React from "react";
-import firebase from "firebase";
-import $ from "jquery"
-
-
-// let firebaseConfig = {
-//     // Your web app's Firebase configuration
-//       apiKey: "AIzaSyBLo8Qodatxni6vT2np4XGvNVgz1-XTHMY",
-//       authDomain: "moviem8s.firebaseapp.com",
-//       databaseURL: "https://moviem8s.firebaseio.com",
-//       projectId: "moviem8s",
-//       storageBucket: "moviem8s.appspot.com",
-//       messagingSenderId: "223030327474",
-//       appId: "1:223030327474:web:8f756c893c3cd008f6f1ab"
-//     };
-
-// firebase.initializeApp(firebaseConfig);
-// let dbFavMovies = firebase.firestore();
-
-
-
-
-function Table(){
-
-    return null 
-    
-}
-
-export default Table;
